InsufficientSamplesError makes estimates
